src/requirements.txt
ahk==0.11.1
aiohttp==3.6.3
aiorun==2020.11.1
altgraph==0.17
amqp==5.0.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
certifi==2020.6.20
chardet==3.0.4
click==6.7
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
decorator==4.4.2
Flask==1.1.2
Flask-Celery==2.4.3
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Flask-SocketIO==4.3.1
future==0.18.2
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
keyboard==0.13.5
kombu==5.0.2
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.19.2
pandas==1.1.3
parso==0.7.1
passlib==1.7.4
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.8
pyahocorasick==1.4.4
Pygments==2.7.2
pygtail==0.11.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
pywin32-ctypes==0.2.0
pyzmq==19.0.2
quakestats==1.4.0
redis==3.5.3
requests==2.24.0
scapy==2.4.4
six==1.15.0
soupsieve==2.0.1
termcolor==1.1.0
traitlets==5.0.5
twitchio==1.1.0
urllib3==1.25.11
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
yarl==1.5.1
